refactor: Break singleton index.ts file
Also adds named WalkableBuffer export (still default)
