+++ src/Encoding.ts
@@ -1,0 +1,1 @@
+export type Encoding = 'ascii' | 'utf8' | 'utf16le' | 'ucs2' | 'base64' | 'hex';
