+++ src/Endianness.ts
@@ -1,0 +1,7 @@
+/**
+ * The endianness to read numbers with.
+ *
+ * * `LE` - for little-endian
+ * * `BE` - for big-endian
+ */
+export type Endianness = 'BE' | 'LE';
